Load all additional/custom components at any type.
