+++ frontend/js/main-form.js
@@ -57,7 +57,6 @@ import permissions from '@/store/modules/permissions'
 // mixins
 import formatPermalink from '@/mixins/formatPermalink'
 import editorMixin from '@/mixins/editor'
-import BlockMixin from '@/mixins/block'
 import retrySubmitMixin from '@/mixins/retrySubmit'
 import cloneDeep from 'lodash/cloneDeep'
 import isEqual from 'lodash/isEqual'
@@ -107,51 +106,6 @@ Vue.component('a17-editor', a17Editor)
 
 // Add attributes
 Vue.component('a17-modal-add', a17ModalAdd)
-
-// Blocks
-const registerBlockComponent = (name, component) => {
-  return !Vue.options.components[name]
-    ? Vue.component(name, component)
-    : false
-}
-
-if (typeof window[process.env.VUE_APP_NAME].TWILL_BLOCKS_COMPONENTS !== 'undefined') {
-  window[process.env.VUE_APP_NAME].TWILL_BLOCKS_COMPONENTS.map(componentName => {
-    return registerBlockComponent(componentName, {
-      template: '#' + componentName,
-      mixins: [BlockMixin]
-    })
-  })
-}
-
-const extractComponentNameFromContextKey = (contextKey) => `a17-${contextKey.match(/\w+/)[0].replace(/([a-z])([A-Z])/g, '$1-$2').replace(/\s+/g, '-').toLowerCase()}`
-
-const importedCustomBlocks = require.context('@/components/blocks/customs/', false, /\.(js|vue)$/i)
-importedCustomBlocks.keys().map(block => {
-  const componentName = extractComponentNameFromContextKey(block.replace(/customs\//, ''))
-  return registerBlockComponent(componentName, importedCustomBlocks(block).default)
-})
-
-const importedTwillBlocks = require.context('@/components/blocks/', false, /\.(js|vue)$/i)
-importedTwillBlocks.keys().map(block => {
-  const componentName = extractComponentNameFromContextKey(block)
-  return registerBlockComponent(componentName, importedTwillBlocks(block).default)
-})
-
-// Custom form components
-const importedComponents = require.context(process.env.VUE_APP_CUSTOM_COMPONENTS_PATH, true, /\.(js|vue)$/i)
-importedComponents.keys().map(block => {
-  // eslint-disable-next-line
-  const componentName = extractComponentNameFromContextKey(block)
-  return Vue.component(componentName, importedComponents(block).default)
-})
-
-// Vendor form components
-const importedVendorComponents = require.context('@/components/customs-vendor/', true, /\.(js|vue)$/i)
-importedVendorComponents.keys().map(block => {
-  const componentName = extractComponentNameFromContextKey(block)
-  return Vue.component(componentName, importedVendorComponents(block).default)
-})
 
 /* eslint-disable no-new */
 /* eslint no-unused-vars: "off" */
